return ids to int type, refactor sql query

+++ src/main/java/com/vlsu/demo/model/CustomUserDetails.java
@@ -7,19 +7,19 @@ import java.util.Collection;
 
 public class CustomUserDetails extends User {
 
-    private int userId;
+    private Long userId;
 
     public CustomUserDetails(String username, String password, boolean enabled, boolean accountNonExpired
-            , boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities, int userId) {
+            , boolean credentialsNonExpired, boolean accountNonLocked, Collection<? extends GrantedAuthority> authorities, Long userId) {
         super(username, password, enabled, accountNonExpired, credentialsNonExpired, accountNonLocked, authorities);
         this.userId = userId;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 }
